# specs/004-map-zoom-annotations/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: MolfMaps API
  version: 0.3.0
  description: >
    REST contract after adding map annotations (feature 004-map-zoom-annotations).
    Only the annotation endpoints are new. The endpoints inherited from 001/002/003
    (/api/map, /api/login, /api/logout, /api/session, /api/maps*) are unchanged and
    are documented in specs/003-maps-list-admin/contracts/openapi.yaml; they are
    referenced but not reproduced here. The authoritative runtime schema remains the
    OpenAPI document generated by FastAPI.

    Authorization model: listing annotations and fetching point-of-interest images
    is public — viewing maps, following text links, and opening popups never require
    a session (FR-048). Creating, editing, resizing, repositioning, and deleting
    annotations, and attaching or detaching images, all require a valid admin session
    and are enforced server-side (FR-045, FR-046). Hiding the placement toggles in the
    UI is presentation only.

    Zoom and pan are entirely client-side view state and have no API surface
    (research Decision 1). Placement mode is likewise client-only and is never sent
    to the server.
servers:
  - url: http://localhost:8000
    description: Local backend (Uvicorn). Angular dev uses a proxy so the browser
      calls same-origin /api/* on the ng serve origin.
paths:
  /api/maps/{map_id}/annotations:
    parameters:
      - $ref: "#/components/parameters/MapId"
    get:
      summary: List a map's annotations
      description: >
        Returns every annotation on the map, ordered by creation time ascending.
        No authentication required — text links and point-of-interest popups are
        visible to every visitor (FR-028, FR-048). A map with no annotations
        returns an empty array and behaves exactly as before (spec edge case).
        Each text link carries `target_available`, computed from the current
        catalog, so the UI can mark a stale link without a lookup per label.
      operationId: listAnnotations
      security: []
      responses:
        "200":
          description: The map's annotations (possibly empty).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnotationListResponse"
        "404":
          description: No map with this id exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create an annotation (admin only)
      description: >
        Creates a text link or a point of interest at a position on the map
        (FR-020, FR-031). The position comes from the click that armed the
        placement mode and is expressed as fractions of the map image, so the
        annotation stays anchored through zoom, pan, resize, and reload
        (FR-038). Requires a valid admin session (FR-045).

        A text link requires both `text` and an existing `target_map_id`;
        incomplete submissions are rejected and nothing is saved (FR-021). A
        point of interest requires `text` and is created with no images —
        images are attached afterwards through the images sub-resource, which
        is why a POI with text but no images is a valid, successful result
        (FR-033).

        `text_scale` is optional and defaults to 0.03 (FR-025); values outside
        the allowed range are clamped to the nearest bound rather than
        rejected (FR-024).
      operationId: createAnnotation
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnotationCreateRequest"
      responses:
        "201":
          description: Annotation created and now visible to every visitor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Annotation"
        "401":
          description: >
            No valid admin session. Annotations are unchanged (FR-046).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No map with this id exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: >
            Unknown kind, missing or over-long text, missing target map, a
            target map that does not exist, or a position outside [0, 1].
            Nothing was saved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/maps/{map_id}/annotations/{annotation_id}:
    parameters:
      - $ref: "#/components/parameters/MapId"
      - $ref: "#/components/parameters/AnnotationId"
    patch:
      summary: Edit, resize, or reposition an annotation (admin only)
      description: >
        Applies a partial update. One endpoint covers all three authoring
        corrections because each is a partial field update on the same record:
        editing wording, target map, or point-of-interest text (FR-039);
        changing a label's text size (FR-022); and moving the annotation to a
        different spot (FR-040). Omitted fields are left unchanged.

        `kind` is immutable — changing a text link into a point of interest
        means deleting it and creating a new one. Fields belonging to the other
        kind are rejected. `text_scale` is clamped to the allowed range rather
        than rejected (FR-024). Successful updates are reflected for every
        visitor on the next read (FR-042).
      operationId: updateAnnotation
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnotationUpdateRequest"
      responses:
        "200":
          description: The updated annotation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Annotation"
        "401":
          description: >
            No valid admin session. The annotation is unchanged (FR-046).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No such map, or no such annotation on this map.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: >
            Empty or over-long text, a target map that does not exist, a
            position outside [0, 1], or a field that does not belong to this
            annotation's kind. Nothing was changed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete an annotation (admin only)
      description: >
        Permanently removes the annotation and, for a point of interest, its
        stored image files (FR-041). The explicit confirmation step required by
        FR-041 is a frontend concern; this endpoint applies the deletion
        immediately when called. The removal is visible to every visitor on the
        next read (FR-042).
      operationId: deleteAnnotation
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Annotation deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: >
            No valid admin session. The annotation remains (FR-046).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No such map, or no such annotation on this map.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/maps/{map_id}/annotations/{annotation_id}/images:
    parameters:
      - $ref: "#/components/parameters/MapId"
      - $ref: "#/components/parameters/AnnotationId"
    post:
      summary: Attach an image to a point of interest (admin only)
      description: >
        Adds one image to the end of the point of interest's ordered image list
        (FR-032). Used by both the create flow and the edit flow, so image
        handling exists once. The image is validated by sniffing its leading
        bytes against the same allowlist and size cap used for map uploads; the
        client-supplied filename and Content-Type are ignored and the stored
        filename derives from a server-generated id (FR-037).

        A refused upload leaves the point of interest unchanged rather than in a
        broken state, which is why a text-only point of interest remains valid
        after a failed attachment.
      operationId: addAnnotationImage
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/AddImageRequest"
      responses:
        "201":
          description: Image stored and appended to the popup's image list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoiImage"
        "400":
          description: >
            The uploaded file is empty, corrupt, or not an allowed image type.
            Nothing was stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No valid admin session. The point of interest is unchanged.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No such map, or no such annotation on this map.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: >
            The annotation is a text link, which carries no images, or the point
            of interest already holds MAX_POI_IMAGES images.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: >
            The image exceeds MAX_MAP_IMAGE_BYTES. The upload is aborted mid
            stream and nothing is stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Missing image part.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/maps/{map_id}/annotations/{annotation_id}/images/{image_id}:
    parameters:
      - $ref: "#/components/parameters/MapId"
      - $ref: "#/components/parameters/AnnotationId"
      - $ref: "#/components/parameters/ImageId"
    get:
      summary: Get a point-of-interest image's bytes
      description: >
        Streams the stored image with its detected media type. No authentication
        required, so logged-out visitors can open popups (FR-048). Returns 404
        when the annotation or its file is missing, which the frontend renders as
        a graceful fallback beside the still-readable text rather than a broken
        image (FR-036).
      operationId: getAnnotationImage
      security: []
      responses:
        "200":
          description: The image.
          content:
            image/webp:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
            image/gif:
              schema:
                type: string
                format: binary
        "404":
          description: Map, annotation, or image unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Detach an image from a point of interest (admin only)
      description: >
        Removes the image from the point of interest's list and unlinks the
        stored file. Part of editing a point of interest's images (FR-039).
      operationId: deleteAnnotationImage
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Image removed from the popup.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: No valid admin session. The images are unchanged.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No such map, annotation, or image.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    MapId:
      name: map_id
      in: path
      required: true
      description: >
        Server-generated map identifier (32 lowercase hex characters). Declared
        as an opaque string at runtime, so an id of any other shape resolves to
        404 rather than a validation error.
      schema:
        type: string
        pattern: "^[0-9a-f]{32}$"
    AnnotationId:
      name: annotation_id
      in: path
      required: true
      description: >
        Server-generated annotation identifier (32 lowercase hex characters).
        Must belong to the map named in the path; an annotation from another map
        resolves to 404.
      schema:
        type: string
        pattern: "^[0-9a-f]{32}$"
    ImageId:
      name: image_id
      in: path
      required: true
      description: >
        Server-generated point-of-interest image identifier (32 lowercase hex
        characters). Must belong to the annotation named in the path.
      schema:
        type: string
        pattern: "^[0-9a-f]{32}$"
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    Annotation:
      description: >
        One annotation on a map. Discriminated on `kind`, so each variant
        declares exactly the fields it needs rather than sharing one model of
        optional fields.
      oneOf:
        - $ref: "#/components/schemas/TextLinkAnnotation"
        - $ref: "#/components/schemas/PoiAnnotation"
      discriminator:
        propertyName: kind
        mapping:
          text_link: "#/components/schemas/TextLinkAnnotation"
          poi: "#/components/schemas/PoiAnnotation"
    AnnotationBase:
      type: object
      required: [id, map_id, kind, x, y, created_at, updated_at]
      properties:
        id:
          type: string
          pattern: "^[0-9a-f]{32}$"
          description: Opaque, immutable identity.
        map_id:
          type: string
          pattern: "^[0-9a-f]{32}$"
          description: The owning map. An annotation belongs to exactly one map.
        kind:
          type: string
          enum: [text_link, poi]
          description: Immutable after creation.
        x:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: >
            Horizontal position as a fraction of the map image's width. Stored
            relative to the image, never in screen pixels, so it stays correct at
            any viewport size and zoom level (FR-038).
        y:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Vertical position as a fraction of the map image's height.
        created_at:
          type: string
          format: date-time
          description: ISO 8601 UTC. Also the list ordering key.
        updated_at:
          type: string
          format: date-time
          description: ISO 8601 UTC. Equal to created_at until the first edit.
    TextLinkAnnotation:
      allOf:
        - $ref: "#/components/schemas/AnnotationBase"
        - type: object
          required: [text, target_map_id, text_scale, target_available]
          properties:
            kind:
              type: string
              const: text_link
            text:
              type: string
              minLength: 1
              maxLength: 120
              description: The displayed label; whitespace is trimmed server-side.
            target_map_id:
              type: string
              pattern: "^[0-9a-f]{32}$"
              description: >
                The map this label opens. May be the owning map itself —
                self-references are allowed and simply keep that map's view.
            text_scale:
              type: number
              format: float
              minimum: 0.01
              maximum: 0.1
              description: >
                Font size as a fraction of the map image's width, so the label
                keeps the same proportion to the map at every zoom level
                (FR-027) and looks identical to every visitor (FR-026).
            target_available:
              type: boolean
              description: >
                Computed at read time, never stored: whether target_map_id still
                exists in the catalog. False means following the link shows the
                "no longer available" message (FR-030).
    PoiAnnotation:
      allOf:
        - $ref: "#/components/schemas/AnnotationBase"
        - type: object
          required: [text, images]
          properties:
            kind:
              type: string
              const: poi
            text:
              type: string
              minLength: 1
              maxLength: 2000
              description: >
                Descriptive text shown in the popup. Required — a point of
                interest is never saved without it (FR-033).
            images:
              type: array
              description: >
                Ordered images shown in the popup. Empty is valid and common
                (FR-033). The count is capped at MAX_POI_IMAGES (default 5) by
                the attach endpoint; because that cap is configuration rather
                than contract, it is enforced there and not declared as
                maxItems here.
              items:
                $ref: "#/components/schemas/PoiImage"
    PoiImage:
      type: object
      required: [id, image_url]
      properties:
        id:
          type: string
          pattern: "^[0-9a-f]{32}$"
        image_url:
          type: string
          description: >
            Relative URL for the image bytes. Clients use this value verbatim
            instead of composing paths.
    AnnotationListResponse:
      type: object
      required: [annotations]
      properties:
        annotations:
          type: array
          description: >
            The map's annotations, oldest first. Empty when the map has none.
            An object rather than a bare array, so fields can be added later
            without a breaking change.
          items:
            $ref: "#/components/schemas/Annotation"
    AnnotationCreateRequest:
      description: Discriminated on `kind`, matching the response union.
      oneOf:
        - $ref: "#/components/schemas/CreateTextLinkRequest"
        - $ref: "#/components/schemas/CreatePoiRequest"
      discriminator:
        propertyName: kind
        mapping:
          text_link: "#/components/schemas/CreateTextLinkRequest"
          poi: "#/components/schemas/CreatePoiRequest"
    CreateTextLinkRequest:
      type: object
      required: [kind, x, y, text, target_map_id]
      additionalProperties: false
      properties:
        kind:
          type: string
          const: text_link
        x:
          type: number
          format: float
          minimum: 0
          maximum: 1
        y:
          type: number
          format: float
          minimum: 0
          maximum: 1
        text:
          type: string
          minLength: 1
          maxLength: 120
        target_map_id:
          type: string
          description: >
            Must reference a map that exists at write time; an unknown id is
            rejected with 422 so the author gets immediate feedback rather than
            a link that was broken from birth.
        text_scale:
          type: number
          format: float
          description: >
            Optional. Defaults to 0.03 when omitted (FR-025). Values outside
            [0.01, 0.10] are clamped to the nearest bound rather than rejected,
            matching "the size stops at that limit" (FR-024).
    CreatePoiRequest:
      type: object
      required: [kind, x, y, text]
      additionalProperties: false
      properties:
        kind:
          type: string
          const: poi
        x:
          type: number
          format: float
          minimum: 0
          maximum: 1
        y:
          type: number
          format: float
          minimum: 0
          maximum: 1
        text:
          type: string
          minLength: 1
          maxLength: 2000
    AnnotationUpdateRequest:
      type: object
      additionalProperties: false
      description: >
        Partial update covering edit, resize, and reposition. Every field is
        optional; omitted fields are left unchanged. `kind` cannot be changed,
        and supplying a field that belongs to the other kind is rejected with
        422.
      properties:
        x:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: New horizontal position (FR-040).
        y:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: New vertical position (FR-040).
        text:
          type: string
          minLength: 1
          maxLength: 2000
          description: >
            New label text or point-of-interest text. The schema declares the
            looser of the two caps; the kind-specific limit (120 for a label,
            2000 for a point of interest) is applied against the annotation's
            own kind and answers 422.
        target_map_id:
          type: string
          description: Text links only. Must reference an existing map.
        text_scale:
          type: number
          format: float
          description: Text links only. Clamped to [0.01, 0.10].
    AddImageRequest:
      type: object
      required: [image]
      properties:
        image:
          type: string
          format: binary
          description: >
            Image file. Must be WebP, PNG, JPEG, or GIF as determined by content
            sniffing, and within MAX_MAP_IMAGE_BYTES — the same handling and
            limits already used for map images.
...
    MessageResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: >
            Generic message safe to show a user. Never reveals internal paths or
            stack details.